+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+
+## 1.18.2.0
+
+* During validation, check the number of columns
+* During validation, check the number of lines
+* Validate train files
+
+## 1.18.1.0
+
+* During validation, check whether the maximum values is obtained with the expected data
+
 ## 1.18.0.0
 
 * Add --validate option